login and navbar hotfix
