+++ app/index.html
@@ -13,9 +13,7 @@
     <nav class="navbar">
         <ul class="nav-list">
             <li class="nav-item"><a href="index.html">Home</a></li>
-            <li class="nav-item">
-                <a id="restaurants-link" href="restaurants/restaurants.html">Restaurants</a>
-              </li>
+            <li class="nav-item" id="restaurants-link"><a href="restaurants/restaurants.html">Restaurants</li>
             <li class="nav-item" id="toursOverview-link"><a
                     href="tours/pages/toursOverview/toursOverview.html?guideId=10">Tour Overview</a></li>
             <li class="nav-item" id="tours-lookup-link"><a href="tours/pages/toursLookup/toursLookup.html">Tour Lookup</li>
