+++ app/index.scss
@@ -15,3 +15,7 @@
 #toursOverview-link {
     display:none;
 }
+
+#tours-lookup-link {
+    display: none
+}
