+++ app/users/pages/login/login.ts
@@ -1,7 +1,6 @@
 import { UserService } from "../../service/user.services.js";
 const userService = new UserService();
 let submitButton: HTMLButtonElement;
-
 document.addEventListener("DOMContentLoaded", () => {
     submitButton = document.getElementById("submit-button") as HTMLButtonElement;
     submitButton.addEventListener("click", (event) => {
@@ -30,6 +29,36 @@ function handleLogin(event: Event) {
             console.error('Login failed', error.message);
         });
 }
+
+export function handlePermission() {
+    const role = localStorage.getItem("role")
+    const loginButton = document.getElementById('login-button') as HTMLElement;
+    const logoutButton = document.getElementById('logout-button') as HTMLElement;
+
+    if (!role) {
+        loginButton.style.display = "flex"
+        logoutButton.style.display = "none"
+        return;
+    }
+
+    if (role == "turista") {
+        const toursLookupLink = document.getElementById("tours-lookup-link") as HTMLLIElement
+        toursLookupLink.style.display = "flex"
+    }
+    else if (role == "vodic") {
+        const toursOverviewLink = document.getElementById("toursOverview-link") as HTMLLIElement
+        toursOverviewLink.style.display = "flex"
+    }
+    else if (role == "vlasnik") {
+        const toursRestaurantsLink = document.getElementById("restaurants-link") as HTMLLIElement
+        toursRestaurantsLink.style.display = "flex"
+    }
+
+    loginButton.style.display = "none"
+    logoutButton.style.display = "flex"
+    logoutButton.addEventListener("click", () => handleLogout())
+}   
+
 
 export function handleLogout() {
     localStorage.removeItem('id');
